requirements.txt
abnf==1.1.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
bitarray==2.7.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
cytoolz==0.12.1
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
fastapi==0.95.1
frozenlist==1.3.3
h11==0.14.0
hexbytes==0.3.0
idna==3.4
jsonschema==4.17.3
lru-dict==1.1.8
multidict==6.0.4
parsimonious==0.9.0
protobuf==4.22.3
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
regex==2023.3.23
requests==2.28.2
rfc3987==1.3.8
rlp==3.0.0
siwe==2.2.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
toolz==0.12.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
web3==6.2.0
websockets==11.0.1
yarl==1.8.2
